refactor: separate SetDepartureTime button
